% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/applyAttributes.R
\name{applyAttributes}
\alias{applyAttributes}
\title{Apply stored attributes}
\usage{
applyAttributes(x, attributesToApply = NULL)
}
\arguments{
\item{x}{a data.frame or input data of class xpssFrame.}

\item{attributesToApply}{to applied attributes}
}
\value{
Object with attributes from \code{\link{attributesBackup}}
}
\description{
Applies attributes stored by attributesBackup
}
\examples{
#load data
data(fromXPSS)
attributes(fromXPSS)
attributes(fromXPSS$V7_2)
x <- attributesBackup(fromXPSS)
fromXPSS <- fromXPSS[order(fromXPSS$V2),]
attributes(fromXPSS)
attributes(fromXPSS$V7_2)
fromXPSS <- applyAttributes(fromXPSS, x)
attributes(fromXPSS)
attributes(fromXPSS$V7_2)
}
\author{
Andreas Wygrabek
}
\seealso{
\code{\link{attributes}} \code{\link{attr}}
}

