% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/applyDateCheck.R
\name{applyDateCheck}
\alias{applyDateCheck}
\title{Apply Date Checks}
\usage{
applyDateCheck(x)
}
\arguments{
\item{x}{a (non-empty) date of class \code{"xpssDate"}.}
}
\value{
xpssDateformat
}
\description{
Checks the date format for inputdata
}
\details{
Helper Function for Merging Attributedatasets
}
\author{
Bastian Wiessner
}
\keyword{internal}

