% Generated by roxygen2 (4.0.2): do not edit by hand
\name{as.xpssFrame}
\alias{as.xpssFrame}
\title{Coerces data.frames to xpssFrame objects}
\usage{
as.xpssFrame(x)
}
\arguments{
\item{x}{any (non-empty) R object.}
}
\description{
Function to coerce a object to xpssFrame
}
\details{
Coerce to xpssFrame.
}
\examples{
# example data.frame
temp <- data.frame(x=1:5, y = 2:6, z=c("a","b","c","d","e"))
# coerce to xpssFrame
temp <- as.xpssFrame(temp)
}
\author{
Bastian Wiessner
}

