% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/xpssComputeArithmetics.R
\name{computeAbs}
\alias{computeAbs}
\title{Computes the absolute value.}
\usage{
computeAbs(x)
}
\arguments{
\item{x}{an atomic numeric or numeric vector or numeric matrix.}
}
\value{
Returns a vector with the absolute values of the input data.
}
\description{
Helper Function for xpssCompute. R Implementation of the SPSS \code{ABS} Function.
}
\examples{
data(fromXPSS)
# atomic numeric input
x1 <- fromXPSS$V5
xpssCompute(x = x1,fun = "computeAbs")
# numeric vector input
x2 <- c(fromXPSS$V7_1,fromXPSS$V7_2,fromXPSS$V5)
xpssCompute(x = x2,fun = "computeAbs")
# numeric matrix input
x3 <- rbind(fromXPSS$V7_2,fromXPSS$V5,fromXPSS$V6)
xpssCompute(x = x3,fun = "computeAbs")
}
\author{
Bastian Wiessner
}
\seealso{
\code{\link{abs}}
}
\keyword{internal}

