% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/xpssComputeArithmetics.R
\name{computeArtan}
\alias{computeArtan}
\title{Computes the arc-tan.}
\usage{
computeArtan(x)
}
\arguments{
\item{x}{an atomic numeric or numeric vector or numeric matrix.}
}
\value{
returns a numeric, numeric vector or matrix with the arc-tan values of the input data.
}
\description{
Helper Function for xpssCompute. R Implementation of the SPSS \code{ARTAN} Function.
}
\examples{
# atomic numeric input
xpssCompute(x = fromXPSS, variables = "V5",fun = "computeArtan")
}
\author{
Bastian Wiessner
}
\seealso{
\code{\link{atan}}
}
\keyword{internal}

