% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/xpssComputeStatistics.R
\name{computeCfvar}
\alias{computeCfvar}
\title{Computes the coefficient of variation}
\usage{
computeCfvar(x,...)
}
\arguments{
\item{x}{atomic numeric or numeric vector or numeric matrix}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Numeric. Returns the coeffiecent of variation.
If the data contains missing values, it is possible to specify an na remove command, this is \code{na.rm}. The default for na.rm is \code{na.rm=F}, if the value get changed to \code{na.rm=T} every existing missing value get omitted.
}
\description{
Helper Function for xpssCompute. R Implementation of the SPSS \code{CFVAR} Function.
}
\examples{
# numeric vector input
xpssCompute(x = fromXPSS, variables = c("V5","V7_1"),fun = "computeCfvar")
}
\author{
Bastian Wiessner
}
\seealso{
\code{\link{max}}
}
\keyword{internal}

