% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/xpssComputeStrings.R
\name{computeChar_length}
\alias{computeChar_length}
\title{Length of a string in characters}
\usage{
computeChar_length (x)
}
\arguments{
\item{x}{atomic character or character vector.}
}
\value{
Numeric. Returns the length of \code{x} in characters, with any trailing blanks removed.
}
\description{
Helper Function for xpssCompute. R Implementation of the SPSS \code{CHAR.LENGTH} Function.
}
\examples{
xpssCompute(x="            please remove trailing blanks",fun="computeChar_length")

xpssCompute(x="please remove trailing blanks",fun="computeChar_length")

xpssCompute(x=c("please","remove","trailing","blanks"),fun="computeChar_length")
}
\author{
Bastian Wiessner
}
\seealso{
\code{\link{str_length}}
}
\keyword{internal}

