% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/xpssComputeStrings.R
\name{computeChar_lpad}
\alias{computeChar_lpad}
\title{Expand strings on the left}
\usage{
computeChar_lpad (x, length = NULL, fill = NULL)
}
\arguments{
\item{x}{atomic character or character vector.}

\item{length}{atomic numeric. Number of characters x is to be filled on the left.}

\item{fill}{atomic character. String which should be append on the left side.}
}
\value{
String, left-padded by length \code{lengthpa}.
}
\description{
Helper Function for xpssCompute. R Implementation of the SPSS \code{CHAR.LPAD} Function.
}
\details{
The value of length represents the number of characters and must be a positive integer. If the optional argument \code{fill} is not specified, \code{x} is padded with blank spaces.
}
\examples{
xpssCompute(x="My Friend",fun="computeChar_lpad",length=15)

xpssCompute(x="My Friend",fun="computeChar_lpad",length=15, fill="Hello ")
}
\author{
Bastian Wiessner
}
\seealso{
\code{\link{paste0}}
}
\keyword{internal}

