% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/xpssComputeStrings.R
\name{computeChar_mblen}
\alias{computeChar_mblen}
\title{Byte per character or sign}
\usage{
computeChar_mblen (x,pos = NULL)
}
\arguments{
\item{x}{character vector, or a vector to be coerced to a character vector. Giving a factor is an error.}

\item{pos}{position of character or sign the number of bytes return to.}
}
\value{
Numeric. Returns the number of bytes in the character at position \code{pos} in \code{x}.
}
\description{
Helper Function for xpssCompute. R Implementation of the SPSS \code{CHAR.MBLEN} Function.
}
\details{
Important for Asian languages, where a character can fill more than one byte.
}
\examples{
#Returns 1 cause the letter "R" fills one byte.
computeChar_mblen("'R' is great!",2)

#Returns 3 cause the asian sign "???" fills three bytes.
computeChar_mblen("???R???????????????????????????!",5)

string <- c("'R' is great!","???R???????????????????????????!","'R'??? ??????!","'R'????????????!")
computeChar_mblen(string,6)
}
\author{
Bastian Wiessner
}
\seealso{
\code{\link{nchar}}
}
\keyword{internal}

