% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/xpssComputeStrings.R
\name{computeChar_substr}
\alias{computeChar_substr}
\title{creates a substring}
\usage{
computeChar_substr (x,pos = NULL, length= NULL)
}
\arguments{
\item{x}{input character vector.}

\item{pos}{atomic numeric. Indicates the start of the substring.}

\item{length}{atomic numeric. Specifies the length of the substring.}
}
\value{
String. Returns an shortened string.
}
\description{
Helper Function for xpssCompute. R Implementation of the SPSS \code{CHAR.SUBSTR} Function.
}
\examples{
xpssCompute(x=fromXPSS, variables="V1", fun="computeChar_substr", pos = 2, length=3)
}
\author{
Bastian Wiessner
}
\seealso{
\code{\link{substr}}
}
\keyword{internal}

