% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/xpssComputeArithmetics.R
\name{computeCos}
\alias{computeCos}
\title{Computes the cosinus.}
\usage{
computeCos(x)
}
\arguments{
\item{x}{an atomic numeric or numeric vector or numeric matrix.}
}
\value{
returns a numeric, numeric vector or matrix with the cosinus values of the input data.
}
\description{
Helper Function for xpssCompute. R Implementation of the SPSS \code{COS} Function.
}
\examples{
# atomic numeric input
xpssCompute(x = fromXPSS, variables = "V5",fun = "computeCos")
}
\author{
Bastian Wiessner
}
\seealso{
\code{\link{cos}}
}
\keyword{internal}

