% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/xpssComputeDates.R
\name{computeDate_dmy}
\alias{computeDate_dmy}
\title{Creates a  date with the format day-month-year}
\usage{
computeDate_dmy(day=NULL,month=NULL, year= NULL)
}
\arguments{
\item{day}{atomic numeric or integer.}

\item{month}{atomic numeric or integer.}

\item{year}{atomic numeric or integer.}
}
\value{
Returns a date object vector of the structure day-month-year.
}
\description{
R Implementation of the SPSS \code{DATE.DMY} Function. \code{computeDate_dmy} is a helper function for xpssCompute.
}
\details{
An character string as Date. Returns a date value corresponding to the indicated day, month, and year. The arguments must resolve to integers or numerics, with day between 1 and 31, month between 1 and 12, and year a four-digit integer value.
}
\examples{
# "05-12-2006"
computeDate_dmy(day=5,month=12,year=2006)
# "10-10-2010"
computeDate_dmy(day=10,month=10,year=2010)
}
\author{
Bastian Wiessner
}
\seealso{
\code{\link{computeDate_mdy}} \code{\link{computeDate_moyr}}
}
\keyword{internal}

