% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/xpssComputeDates.R
\name{computeDate_mdy}
\alias{computeDate_mdy}
\title{Creates a date with the format month-day-year}
\usage{
computeDate_mdy(month=NULL,day=NULL, year= NULL)
}
\arguments{
\item{month}{atomic numeric or integer.}

\item{day}{atomic numeric or integer.}

\item{year}{atomic numeric or integer.}
}
\value{
Returns a character or character vector of the structure month-day-year
}
\description{
R Implementation of the SPSS \code{DATE.MDY} Function. \code{computeDate_mdy} is a helper function for xpssCompute.
}
\details{
An character string as Date. Returns a date value corresponding to the indicated day, month, and year. The arguments must resolve to integers or numerics, with day between 1 and 31, month between 1 and 12, and year a four-digit integer value.
}
\examples{
# "05-12-2006"
xpssCompute(x=5, fun="computeDate_dmy", day=12,year=2006)
# "10-10-2010"
xpssCompute(x=10, fun="computeDate_dmy", day=10,year=2010)
}
\author{
Bastian Wiessner
}
\seealso{
\code{\link{computeDate_dmy}} \code{\link{computeDate_moyr}}
}
\keyword{internal}

