% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/xpssComputeDates.R
\name{computeDate_moyr}
\alias{computeDate_moyr}
\title{Creates a date with the format month-year}
\usage{
computeDate_moyr(month=NULL, year= NULL)
}
\arguments{
\item{month}{atomic numeric or integer.}

\item{year}{atomic numeric or integer.}
}
\value{
Returns a character or character vector of the structure month-year.
}
\description{
R Implementation of the SPSS \code{DATE.MOYR} Function. \code{computeDate_moyr} is a helper function for xpssCompute.
}
\details{
An character string as Date. Returns a date value corresponding to the indicated month, and year. The arguments must resolve to integers or numerics, with month between 1 and 12, and year a four-digit integer value.
}
\examples{
# 10-2012
computeDate_moyr(month=10, year=2012)
# 05-1999
computeDate_moyr(month=5, year=1999)
}
\author{
Bastian Wiessner
}
\seealso{
\code{\link{computeDate_dmy}} \code{\link{computeDate_mdy}}
}
\keyword{internal}

