% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/xpssComputeArithmetics.R
\name{computeExp}
\alias{computeExp}
\title{Computes the exponential}
\usage{
computeExp(x)
}
\arguments{
\item{x}{an atomic numeric or numeric vector or numeric matrix.}
}
\value{
returns a numeric, numeric vector or matrix with the expontential values of the input data.
}
\description{
Helper Function for xpssCompute. R Implementation of the SPSS \code{EXP} Function.
}
\examples{
xpssCompute(x = fromXPSS, variables = "V5",fun = "computeExp")
}
\author{
Bastian Wiessner
}
\seealso{
\code{\link{exp}}
}
\keyword{internal}

