% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/xpssComputeStrings.R
\name{computeLength}
\alias{computeLength}
\title{Number of bytes in a string}
\usage{
computeLength (x)
}
\arguments{
\item{x}{character or character vector.}
}
\value{
Numeric. Returns the length of \code{x} in bytes, including all trailing blanks.
}
\description{
Helper Function for xpssCompute. R Implementation of the SPSS \code{LENGTH} Function.
}
\examples{
computeLength("trailing blanks matter")
computeLength("trailing blanks matter          ")
computeLength("        trailing blanks matter")
}
\author{
Bastian Wiessner
}
\seealso{
\code{\link{nchar}}
}
\keyword{internal}

