% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/xpssComputeArithmetics.R
\name{computeLg10}
\alias{computeLg10}
\title{Computes the logarithm base 10}
\usage{
computeLg10(x)
}
\arguments{
\item{x}{an atomic numeric or numeric vector or numeric matrix.}
}
\value{
returns a numeric, numeric vector or matrix with the logarithm base 10 values of the input data.
}
\description{
Helper Function for xpssCompute. R Implementation of the SPSS \code{LG10} Function.
}
\examples{
# atomic numeric input
xpssCompute(x = fromXPSS, variables = "V5",fun = "computeLg10")
}
\author{
Bastian Wiessner
}
\seealso{
\code{\link{log}}
}
\keyword{internal}

