% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/xpssComputeArithmetics.R
\name{computeLngamma}
\alias{computeLngamma}
\title{Computes the logarithm of the gamma function}
\usage{
computeLngamma(x)
}
\arguments{
\item{x}{an atomic numeric or numeric vector or numeric matrix.}
}
\value{
returns a numeric, numeric vector or matrix with the logarithm of the gamma function values of the input data.
}
\description{
Helper Function for xpssCompute. R Implementation of the SPSS \code{LNGAMMA} Function.
}
\examples{
# atomic numeric input
xpssCompute(x = fromXPSS, variables = "V5",fun = "computeLngamma")
}
\author{
Bastian Wiessner
}
\seealso{
\code{\link{lgamma}}
}
\keyword{internal}

