% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/xpssComputeStrings.R
\name{computeLower}
\alias{computeLower}
\title{Convert upper-case letters to lower-case}
\usage{
computeLower (x)
}
\arguments{
\item{x}{a character vector, or an object that can be coerced to character by as.character.}
}
\value{
String. Returns unput with uppercase letters changed to lowercase. The argument can be a string variable or a value.
}
\description{
Helper Function for xpssCompute. R Implementation of the SPSS \code{LOWER} Function.
}
\examples{
computeLower("MAKE ME small PLEASE")

computeLower("From BIG 2 small")
}
\author{
Bastian Wiessner
}
\seealso{
\code{\link{tolower}}
}
\keyword{internal}

