% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/xpssComputeStrings.R
\name{computeLtrim}
\alias{computeLtrim}
\title{Trims string on left side}
\usage{
computeLtrim (x,trim = NULL)
}
\arguments{
\item{x}{input character vector}

\item{trim}{single quoted character or an expression that yields a single character.}
}
\value{
String. Returns the input string removed by \code{trim} on the left. If trim is not specified, leading blanks are removed.
}
\description{
Helper Function for xpssCompute. R Implementation of the SPSS \code{LTRIM} Function.
}
\examples{
xpssCompute(x="           inputstring", fun="computeLtrim")
}
\author{
Bastian Wiessner
}
\seealso{
\code{\link{str_trim}}
}
\keyword{internal}

