% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/xpssComputeStatistics.R
\name{computeMean}
\alias{computeMean}
\title{Atithmetic Mean}
\usage{
computeMean (x,...)
}
\arguments{
\item{x}{Numeric or numeric vector}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Numeric. Returns the arithmetic mean of its arguments that have valid, nonmissing values.
If the data contains missing values, it is possible to specify an na remove command, this is \code{na.rm}. The default for na.rm is \code{na.rm=F}, if the value get changed to \code{na.rm=T} every existing missing value get omitted.
}
\description{
Helper Function for xpssCompute. R Implementation of the SPSS \code{MEAN} Function.
}
\examples{
data(fromXPSS)
# atomic numeric input
x1 <- fromXPSS$V5
xpssCompute(x = x1,fun = "computeMean")
# numeric vector input
x2 <- c(fromXPSS$V7_1,fromXPSS$V7_2,fromXPSS$V5)
# optional with na remove
xpssCompute(x = x2,fun = "computeMean",na.rm=T)
# numeric matrix input, output is NA
x3 <- rbind(fromXPSS$V7_2,fromXPSS$V4)
xpssCompute(x = x3,fun = "computeMax")
# numeric matrix input, output is Mean
x4 <- rbind(fromXPSS$V7_2,fromXPSS$V4)
xpssCompute(x = x3,fun = "computeMax", na.rm = T)
}
\author{
Bastian Wiessner
}
\seealso{
\code{\link{mean}}
}
\keyword{internal}

