% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/xpssComputeStatistics.R
\name{computeMedian}
\alias{computeMedian}
\title{Median Value}
\usage{
computeMedian (x,...)
}
\arguments{
\item{x}{Numeric or numeric vector}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Numeric. Returns the median (50th percentile) of its arguments that have valid, nonmissing values.
If the data contains missing values, it is possible to specify an na remove command, this is \code{na.rm}. The default for na.rm is \code{na.rm=F}, if the value get changed to \code{na.rm=T} every existing missing value get omitted.
}
\description{
Helper Function for xpssCompute. R Implementation of the SPSS \code{MEDIAN} Function.
}
\examples{
# numeric vector input
xpssCompute(x = fromXPSS, variables = c("V5","V7_1"),fun = "computeMedian")
}
\author{
Bastian Wiessner
}
\seealso{
\code{\link{median}}
}
\keyword{internal}

