% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/xpssComputeStatistics.R
\name{computeMin}
\alias{computeMin}
\title{Minima}
\usage{
computeMin (x,...)
}
\arguments{
\item{x}{Numeric or numeric vector}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Numeric or string. Returns the minimum value of its arguments that have valid, nonmissing values.
If the data contains missing values, it is possible to specify an na remove command, this is \code{na.rm}. The default for na.rm is \code{na.rm=F}, if the value get changed to \code{na.rm=T} every existing missing value get omitted.
}
\description{
Helper Function for xpssCompute. R Implementation of the SPSS \code{MIN} Function.
}
\examples{
# numeric vector input
xpssCompute(x = fromXPSS, variables = c("V5","V7_1"),fun = "computeMin")
}
\author{
Bastian Wiessner
}
\seealso{
\code{\link{min}}
}
\keyword{internal}

