% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/xpssComputeArithmetics.R
\name{computeMod}
\alias{computeMod}
\title{Returns the remainder of a division.}
\usage{
computeMod(x,modulus = NULL)
}
\arguments{
\item{x}{an atomic numeric or numeric vector or numeric matrix.}

\item{modulus}{atomic numeric \code{x} is divided by.}
}
\value{
returns a numeric vector including the remainder of \code{x} by \code{modulus}.
}
\description{
Helper Function for xpssCompute. R Implementation of the SPSS \code{MOD} Function.
}
\examples{
# atomic numeric input
xpssCompute(x = fromXPSS, variables = "V5",fun = "computeMod", modulus = 2)
}
\author{
Bastian Wiessner
}
\seealso{
\code{\link{\%\%}}
}
\keyword{internal}

