% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/xpssComputeStrings.R
\name{computeNtrim}
\alias{computeNtrim}
\title{Return the input data without removing trailing blanks}
\usage{
computeNtrim (x)
}
\arguments{
\item{x}{input character vector}
}
\value{
String. Returns the values of the input data
}
\description{
Helper Function for xpssCompute. R Implementation of the SPSS \code{NTRM} Function.
}
\examples{
xpssCompute(x=fromXPSS,variables="V1", fun="computeNtrim")
}
\author{
Bastian Wiessner
}
\keyword{internal}

