% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/xpssComputeArithmetics.R
\name{computeRnd}
\alias{computeRnd}
\title{Rounds values.}
\usage{
computeRnd(x, digits=2)
}
\arguments{
\item{x}{a numeric or complex vector.}

\item{digits}{atomic numeric argument, indicating the number of decimal places.}
}
\value{
returns the rounded values of \code{x} with the specified number of decimal places.
}
\description{
Helper Function for xpssCompute. R Implementation of the SPSS \code{RND} Function.
}
\examples{
# atomic numeric input
x1 <- seq(from = 1,to = 20,by = 1.5)
xpssCompute(x = x1,fun = "computeRnd", digits = 2)
# numeric vector input
x2 <- c(seq(from = 1,to = 20,by = 1.5),seq(from = 1,to = 20,by = 1.25))
xpssCompute(x = x2,fun = "computeRnd", digits = 3)
# numeric matrix input
x3 <- rbind(seq(from = 1,to = 10,by = 1),seq(from = 1,to = 20,by = 2),seq(from = 1,to = 30,by = 3))
xpssCompute(x = x3,fun = "computeRnd", digits = 1)
}
\author{
Bastian Wiessner
}
\seealso{
\code{\link{round}}
}
\keyword{internal}

