% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/xpssComputeStrings.R
\name{computeRtrim}
\alias{computeRtrim}
\title{Trims string on right side}
\usage{
computeRtrim (x,trim = NULL)
}
\arguments{
\item{x}{input character vector}

\item{trim}{single quoted character or an expression that yields a single character.}
}
\value{
String. Returns the input string removed by \code{trim} on the right. If trim is not specified, trailing blanks are removed.
}
\description{
Helper Function for xpssCompute. R Implementation of the SPSS \code{RTRIM} Function.
}
\examples{
\dontrun{beispiel <- beispiel()}
}
\author{
Bastian Wiessner
}
\seealso{
\code{\link{str_trim}}
}
\keyword{internal}

