% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/xpssComputeArithmetics.R
\name{computeSqrt}
\alias{computeSqrt}
\title{Computes the square root}
\usage{
computeSqrt(x)
}
\arguments{
\item{x}{an atomic numeric or numeric vector or numeric matrix.}
}
\value{
returns a numeric, numeric vector or matrix with the square root values of the input data.
}
\description{
Helper Function for xpssCompute. R Implementation of the SPSS \code{SQRT} Function.
}
\examples{
data(fromXPSS)
# atomic numeric input
x1 <- fromXPSS$V5
xpssCompute(x = x1,fun = "computeSqrt")
# numeric vector input
x2 <- c(fromXPSS$V7_1,fromXPSS$V7_2,fromXPSS$V5)
xpssCompute(x = x2,fun = "computeSqrt")
# numeric matrix input
x3 <- rbind(fromXPSS$V7_2,fromXPSS$V5,fromXPSS$V6)
xpssCompute(x = x3,fun = "computeSqrt")
}
\author{
Bastian Wiessner
}
\seealso{
\code{\link{sqrt}}
}
\keyword{internal}

