% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/xpssComputeStrings.R
\name{computeStrunc}
\alias{computeStrunc}
\title{Truncating strings}
\usage{
computeStrunc (x, length = NULL)
}
\arguments{
\item{x}{input character vector.}

\item{length}{length x is truncated to}
}
\value{
String. Returns \code{x} truncated to \code{length} (in bytes) and trimmed of remaining blanks.
}
\description{
Helper Function for xpssCompute. R Implementation of the SPSS \code{STRUNC} Function.
}
\examples{
computeStrunc("Hello    ",7)
}
\author{
Bastian Wiessner
}
\seealso{
\code{\link{str_sub}}
}
\keyword{internal}

