% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/xpssComputeDates.R
\name{computeTime_days}
\alias{computeTime_days}
\title{Calculates the number of passed hours on basis of the given days}
\usage{
computeTime_days(x=NULL)
}
\arguments{
\item{x}{atomic numeric or integer.}
}
\value{
Returns the number of passed hours as atomic numeric.
}
\description{
R Implementation of the SPSS \code{TIME.DAYS} Function. \code{computeTime_days} is a helper function for xpssCompute.
}
\examples{
# 120
xpssCompute(x=5, fun="computeTime_days")
# 288
xpssCompute(x=12, fun="computeTime_days")
}
\author{
Bastian Wiessner
}
\seealso{
\code{\link{computeDate_qyr}} \code{\link{computeDate_wkyr}}
}
\keyword{internal}

