% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/xpssComputeDates.R
\name{computeTime_hms}
\alias{computeTime_hms}
\title{Creates a date with the format Hour-Minute-Second}
\usage{
computeTime_hms(hour=NULL,minute=NULL,second=NULL)
}
\arguments{
\item{hour}{atomic numeric or integer.}

\item{minute}{atomic numeric or integer.}

\item{second}{atomic numeric or integer.}
}
\value{
Returns a character string as Timedate.
}
\description{
R Implementation of the SPSS \code{TIME.HMS} Function. \code{computeTime_hms} is a helper function for xpssCompute.
}
\examples{
# 4:10:23
computeTime_hms(hour=4,minute=10,second=23)
# 14:34:01
computeTime_hms(hour=14,minute=34,second=1)
}
\author{
Bastian Wiessner
}
\seealso{
\code{\link{computeDate_qyr}} \code{\link{computeDate_wkyr}}
}
\keyword{internal}

