% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/xpssComputeStrings.R
\name{computeUpcase}
\alias{computeUpcase}
\title{Convert lower-case letters to upper-case}
\usage{
computeUpcase (x)
}
\arguments{
\item{x}{a character vector, or an object that can be coerced to character by as.character.}
}
\value{
String. Returns unput with lowercase letters changed to uppercase. The argument can be a string variable or a value.
}
\description{
Helper Function for xpssCompute. R Implementation of the SPSS \code{UPCASE} Function.
}
\examples{
computeUpcase("make it big")

small <- c("make","all","big")
computeUpcase(small)
}
\author{
Bastian Wiessner
}
\seealso{
\code{\link{toupper}}
}
\keyword{internal}

