% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/xpssComputeDates.R
\name{computeXdate_date}
\alias{computeXdate_date}
\title{Extracts the date out of a date string.}
\usage{
computeXdate_date(x=NULL)
}
\arguments{
\item{x}{atomic object of class \code{character, POSIXlt  or  POSIXt} holding date data.}
}
\value{
Returns a character string with date values.
}
\description{
R Implementation of the SPSS \code{XDATE.DATE} Function. \code{computeXdate_date} is a helper function for xpssCompute.
}
\details{
computeXdate_date extract the complete date out of the date string, the time componentes will be omitted.
}
\examples{
# 2015-02-15
computeXdate_date("2015-02-15 20:22:20")
# 2022-02-20
computeXdate_date("2022-02-20 21:22:12")
}
\author{
Bastian Wiessner
}
\seealso{
\code{\link{computeXdate_hour}} \code{\link{computeDate_wkyr}}
}
\keyword{internal}

