% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/xpssComputeDates.R
\name{computeXdate_jday}
\alias{computeXdate_jday}
\title{Calculates the date of year on basis of a given date}
\usage{
computeXdate_jday(x=NULL)
}
\arguments{
\item{x}{atomic object of class \code{character, POSIXlt  or  POSIXt} holding date data.}
}
\value{
Returns a character string with the date of year.
}
\description{
R Implementation of the SPSS \code{XDATE.JDAY} Function. \code{computeXdate_jday} is a helper function for xpssCompute.
}
\details{
computeXdate_jday caluclates the date of year out of the date string, the other date and time componentes will be omitted.
}
\examples{
# 046
computeXdate_jday("2015-02-15 20:22:20")
# 293
computeXdate_jday("2022-10-20 21:22:12")
}
\author{
Bastian Wiessner
}
\seealso{
\code{\link{computeXdate_date}} \code{\link{computeDate_wkyr}}
}
\keyword{internal}

