% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/xpssComputeDates.R
\name{computeXdate_time}
\alias{computeXdate_time}
\title{Extracts the time componente on basis of a given date}
\usage{
computeXdate_time(x=NULL)
}
\arguments{
\item{x}{atomic object of class \code{character, POSIXlt  or  POSIXt} holding date data.}
}
\value{
Returns a character string with the time components.
}
\description{
R Implementation of the SPSS \code{XDATE.TIME} Function. \code{computeXdate_time} is a helper function for xpssCompute.
}
\details{
computeXdate_time extract the time components out of the date string, the other date components will be omitted.
}
\examples{
# 20:11:20
xpssCompute(x="2015-02-15 20:11:20", fun="computeXdate_time")
# 21:22:12
xpssCompute(x="2022-10-20 21:22:12", fun="computeXdate_time")
}
\author{
Bastian Wiessner
}
\seealso{
\code{\link{computeXdate_date}} \code{\link{computeDate_wkyr}}
}
\keyword{internal}

