% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/xpssComputeDates.R
\name{computeXdate_week}
\alias{computeXdate_week}
\title{Calcualtes the calendar week  on basis of a given date}
\usage{
computeXdate_week(x=NULL)
}
\arguments{
\item{x}{atomic object of class \code{character, POSIXlt  or  POSIXt} holding date data.}
}
\value{
Returns a character string with the calendar week.
}
\description{
R Implementation of the SPSS \code{XDATE.WEEK} Function. \code{computeXdate_week} is a helper function for xpssCompute.
}
\details{
computeXdate_week caluclates the calendar week on basis of the given date string.
}
\examples{
# 06
xpssCompute(x="2015-02-15 20:11:20", fun="computeXdate_week")
# 42
xpssCompute(x="2022-10-20 21:22:12", fun="computeXdate_week")
}
\author{
Bastian Wiessner
}
\seealso{
\code{\link{computeXdate_date}} \code{\link{computeDate_wkyr}}
}
\keyword{internal}

