% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/xpssComputeDates.R
\name{computeXdate_year}
\alias{computeXdate_year}
\title{Extracts the year on basis of a given date}
\usage{
computeXdate_year(x=NULL)
}
\arguments{
\item{x}{atomic object of class \code{character, POSIXlt  or  POSIXt} holding date data.}
}
\value{
Returns a character string with the day of week.
}
\description{
R Implementation of the SPSS \code{XDATE.YEAR} Function. \code{computeXdate_year} is a helper function for xpssCompute.
}
\details{
computeXdate_year extracts the year on basis of the given date string.
}
\examples{
# 2015
xpssCompute(x="2015-02-15 20:11:20",fun="computeXdate_year")
# 2022
xpssCompute(x="2022-10-20 21:22:12",fun="computeXdate_year")
}
\author{
Bastian Wiessner
}
\seealso{
\code{\link{computeXdate_date}} \code{\link{computeXdate_wkday}}
}
\keyword{internal}

