% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/xpssComputeDates.R
\name{computeXweek}
\alias{computeXweek}
\title{Calcualtes the calendar week  on basis of a given date}
\usage{
computeXweek(x=NULL)
}
\arguments{
\item{x}{atomic object of class \code{character, POSIXlt  or  POSIXt} holding date data.}
}
\value{
Returns a character string with the calendar week.
}
\description{
R Implementation of the SPSS \code{XDATE.WEEK} Function. \code{computeXweek} is a helper function for xpssCompute.
}
\details{
computeXweek caluclates the calendar week on basis of the given date string.
}
\examples{
# 06
computeXweek("2015-02-15 20:11:20")
# 42
computeXweek("2022-10-20 21:22:12")
}
\author{
Bastian Wiessner
}
\seealso{
\code{\link{computeXdate_date}} \code{\link{computeDate_wkyr}}
}
\keyword{internal}

