% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/xpssComputeDates.R
\name{computeXwkday}
\alias{computeXwkday}
\title{Calcualtes the day of week on basis of a given date}
\usage{
computeXwkday(x=NULL)
}
\arguments{
\item{x}{atomic object of class \code{character, POSIXlt  or  POSIXt} holding date data.}
}
\value{
Returns a character string with the day of week.
}
\description{
R Implementation of the SPSS \code{XDATE.WEEK} Function. \code{computeXwkday} is a helper function for xpssCompute.
}
\details{
computeXwkday caluclates the calendar week on basis of the given date string. Result is a number between 0 and 6. 0 stands for Sunday, 6 for Saturday.
}
\examples{
# 0
computeXwkday("2015-02-15 20:11:20")
# 6
computeXwkday("2022-10-20 21:22:12")
}
\author{
Bastian Wiessner
}
\seealso{
\code{\link{computeXdate_date}} \code{\link{computeDate_wkyr}}
}
\keyword{internal}

