% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/xpssComputeDates.R
\name{computeXyear}
\alias{computeXyear}
\title{Extracts the year on basis of a given date}
\usage{
computeXyear(x=NULL)
}
\arguments{
\item{x}{atomic object of class \code{character, POSIXlt  or  POSIXt} holding date data.}
}
\value{
Returns a character string with the day of week.
}
\description{
R Implementation of the SPSS \code{XDATE.YEAR} Function. \code{computeXyear} is a helper function for xpssCompute.
}
\details{
computeXyear extracts the year on basis of the given date string.
}
\examples{
# 2015
computeXyear("2015-02-15 20:11:20")
# 2022
computeXyear("2022-10-20 21:22:12")
}
\author{
Bastian Wiessner
}
\seealso{
\code{\link{computeXdate_date}} \code{\link{computeDate_wkyr}}
}
\keyword{internal}

