% Generated by roxygen2 (4.0.2): do not edit by hand
\name{enter}
\alias{enter}
\title{enter}
\usage{
enter(...)
}
\arguments{
\item{...}{one or more Variables specified as characters}
}
\value{
returns a list of class summary.lm
}
\description{
Helper Function for xpssRegression
}
\details{
Regression Method. All stated variables are used at once in the Regression Modell.
This function only works inside xpssRegression \code{\link{xpssRegression}}
}
\examples{
\dontrun{
enter("var1","var2","var3")
}
}
\author{
Martin Schneider
}
\seealso{
xpssRegression \code{\link{xpssRegression}}
}
\keyword{internal}

