\name{foo}
\alias{foo}
\docType{data}
\title{
Sample dataset foo
}
\description{
A dataset containg a sequence starting from 1 to 15.
}
\usage{data(foo)}
\format{
  A matrix with 3 rows and 5 variables
}
\details{
 Each variable contains 3 values, A (1,2,3); B (4,5,6); and so on, ... E(13,14,15). 
}
\examples{
data(foo)
}
\keyword{datasets}

