% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/translate2R.R
\docType{data}
\name{kids}
\alias{kids}
\title{Sample dataset}
\format{A data.frame with 9 rows and 10 variables.}
\usage{
data(kids)
}
\description{
Dataset kids
}
\details{
\itemize{
\item{id} {ID. (1-9)}
\item{famid} {Familiy ID (1-3)}
\item{kidname} {Names. ("Al", "Andy",...)}
\item{birth} {Chronological order. (1-3)}
\item{age} {Age. (2-9)}
\item{wt} {Weight. (10-40)}
\item{sex} {Sex. (1,2)}
\item{famname} {Familiy name. ("Hansen", "Petersen", "Lange")}
\item{street} {Street .name ("Blumenstr", "Steinweg", "Forstweg")}
\item{birth_2} {Chronological order. ("first", "second", "third")}
}
}
\keyword{datasets}

