% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/xpssComputeMissings.R
\name{nmiss}
\alias{nmiss}
\title{Displays the amount of missing values in variables.}
\usage{
nmiss(x, variables = NULL)
}
\arguments{
\item{x}{a (non-empty) data.frame, data.table object or input data of class \code{"xpssFrame"}.}

\item{variables}{atomic character or character vector with the names of the variables.}
}
\value{
atomic numeric with the length of the data. Returns the amount of
system- or user-defined missing values of the variables.
}
\description{
R implementation of the SPSS \code{Nmiss} function
}
\details{
Performs a missing value operation. \code{nmiss} displays the amount of system- and user-defined missing values of the variables. \cr User-defined and system-defind missings values get handled as one type of missing value.
}
\examples{
# load data
data(fromXPSS)

# display the amount of user- and system-defined missing values in variable V6 and V7_2
nmiss(fromXPSS, variables=c("V6","V7_2"))
}
\author{
Bastian Wiessner
}
\seealso{
Related Functions \code{\link{missing}} , \code{\link{nvalid}} , \code{\link{sysmis}} ,\code{\link{value}}
}
\keyword{internal}

