% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/xpssComputeMissings.R
\name{sysmis}
\alias{sysmis}
\title{Displays if a variable contains system-missing values}
\usage{
sysmis (x, variable = NULL)
}
\arguments{
\item{x}{a (non-empty) data.frame, data.table object or input data of class "xpssFrame".}

\item{variable}{atomic character with the name of the variable.}
}
\value{
atomic numeric with the length of the data. Returns the amount of system missing values of the variable.
}
\description{
R implementation of the SPSS \code{Sysmis} Function
}
\details{
Performs a missing value operation. Displays only the amount of system missing values of the variable instead of looking after all missing values only system-missings get inspected.

\code{variable} is limited to one variable.
}
\examples{
# load data
data(fromXPSS)

# locate sysmis in variable V7_2
sysmis(fromXPSS, variable="V7_2")
}
\author{
Bastian Wiessner
}
\seealso{
Related Functions \code{\link{missing}} , \code{\link{nmiss}} , \code{\link{nvalid}} ,\code{\link{value}}
}
\keyword{internal}

