% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/translate2R.R
\docType{package}
\name{translateSPSS2R-package}
\alias{translateSPSS2R-package}
\title{Documentation}
\arguments{
\item{Package:}{translateSPSS2R}

\item{Type:}{Package}

\item{Version:}{1.0.0}

\item{Date:}{2015-23-06}

\item{Imports:}{car ,data.table, e1071, foreign, Hmisc, plyr, stringr, tidyr, zoo}

\item{License:}{GPL-3}

\item{Maintainer:}{Andreas Wygrabek}
}
\description{
translateSPSS2R provides a set of functions with translated SPSS commands. The usage is oriented on the handling of the SPSS-Syntax.
}
\details{
Mainly the package has two purposes:
\enumerate{
\item It facilitates SPSS-Users to change over to R.
\item It facilitates migration projects from SPSS to R.
}. For automatic translation a webtool (\url{http://www.eoda.de/en/translate2R.html}) is provided by eoda GmbH. The tool translates SPSS-Syntax to R-Code by the use of translateSPSS2R functions.
}
\author{
Andreas Wygrabek <Andreas.Wygrabek@eoda.de> and Bastian Wiessner <Bastian.Wiessner@eoda.de>
}
\references{
\url{https://service.eoda.de/translater/?lang=en}
\cr \url{http://www.eoda.de/en}
}

