% Generated by roxygen2 (4.0.2): do not edit by hand
\name{xpssCasenum}
\alias{xpssCasenum}
\title{Sequential numbering of cases}
\usage{
xpssCasenum(x)
}
\arguments{
\item{x}{a (non-empty) data.frame or input data of class \code{"xpssFrame"}.}
}
\value{
Returns an sequential atomic numeric or numeric vector.
}
\description{
R Implementation of the SPSS \code{$CASENUM} system variable.
}
\details{
The xpssCasenum counts the number of cases within a variable, respectively the number of observations in the dataset.

This statement can be used as ID-Variable.
}
\examples{
data(fromXPSS)
fromXPSS$id <- xpssCasenum(fromXPSS)
}
\author{
Bastian Wiessner
}
\seealso{
Related Functions \code{\link{xpssDate}} , \code{\link{xpssDate11}}
}

