\name{xpssDate}
\alias{xpssDate}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ 
Date function with two-digit-year.
}
\description{
R Implementation of the SPSS \code{$DATE} system variable
}
\usage{
xpssDate()
}
%- maybe also 'usage' for other objects documented here.
\details{
Current date in international date format with two-digit year
}
\value{
%%  ~Describe the value returned
%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
Bastian Wiessner
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
Related Functions \code{\link{xpssTime}} \code{\link{xpssDate11}} code{\link{xpssJDate}}
}
\examples{
xpssDate()
}
