% Generated by roxygen2 (4.0.2): do not edit by hand
\name{xpssDate11}
\alias{xpssDate11}
\title{Todays date}
\usage{
xpssDate11()
}
\value{
Current date as "dd-mmm-yyyy". Day is a numeric with the length two, month is a character string with the length three and year is a numeric value with the length four.
}
\description{
R implementation of the SPSS \code{$date11} system variable.
}
\details{
Provides the SPSS format of dates
}
\examples{
xpssDate11()
}
\author{
Bastia Wiessner
}

