% Generated by roxygen2 (4.0.2): do not edit by hand
\name{xpssDoIf}
\alias{xpssDoIf}
\title{Creates a subset of cases}
\usage{
xpssDoIf(x, cond = NULL)
}
\arguments{
\item{x}{a (non-empty) data.frame or input data of class \code{"xpssFrame"}.}

\item{cond}{logical expression for subsetting the data}
}
\value{
Output is a subset of the actual dataset under the condition of the logical expression.
}
\description{
xpssDoIf selects cases for analysis based on one or more logical conditions.
}
\details{
The conditions to select cases are specified in a logical expression. These logical expressions can contain relational operators, logical operators and arithmetic operations. \cr \cr

The data is subsetted until \code{\link{xpssEndIf}} restores the original data. All changes made at the subsetted data will be taken over, the excluded data will be untouched!\cr \cr

In a different way to SPSS. Not only data management functions like \code{\link{xpssRecode}} can be used within \code{xpssDoIf}, it is possible to use statistical and descriptiv functions like xpssFrequencies too. \cr \cr
 \strong{NOTE:} For temporary case selection, specify \code{\link{xpssTemporary}} before \code{\link{xpssDoIf}}.
}
\examples{
data(fromXPSS)
# Select all cases which match 1 in V3
temp <- xpssDoIf(x=fromXPSS, cond = "V3 == 1")
# Recode all selected cases within V5
temp <- xpssRecode(x=temp,varin="V5",rec="lo:78 = 1; else = 2")
# End DoIf Subsetting, restore dataset with applied changes
temp <- xpssEndIf(x=temp)
}
\author{
Andreas Wygrabek
}
\seealso{
Related Functions \code{\link{xpssEndIf}}, \code{\link{xpssFilter}}, \code{\link{xpssSelectIf}}, \code{\link{xpssTemporary}}
}

