% Generated by roxygen2 (4.0.2): do not edit by hand
\name{xpssFilter}
\alias{xpssFilter}
\title{Creates a subset without deleting cases}
\usage{
xpssFilter(x,variable = NULL, filtervalue = 1)
}
\arguments{
\item{x}{a (non-empty) data.frame, data.table object or input data of class \code{"xpssFrame"}.}

\item{variable}{one or or more variables given as character.}

\item{filtervalue}{atomic character or atomic numeric which contains the filtervalue.}
}
\value{
Output is a subset of the actual dataset under the predetermined condition of the filtervalue.
}
\description{
R Implementation of the SPSS \code{filter} argument
}
\details{
xpssFilter creates a subset of the actual dataset, without deleting the excluded variables, respectively without deleting the excluded values. After activating \code{\link{xpssFilter}}, only the subset will be used, the excluded data gets ignored for the following actions until  \code{\link{xpssFilterOff}} terminates the filtering. As noted before those cases are \strong{not actually} deleted and will be available after the filter is turned off.
\cr\cr \strong{Important:}
\cr All changes are used on the complete dataset, except the function is a \emph{data exploring} or \emph{data analyzing} function \cr \cr
 \tabular{rlll}{
 \tab Type of Function \tab Example Function \tab Dataset Usage \cr
\tab Data Management  \tab \code{\link{xpssSelectIf}} \tab Uses complete dataset\cr
\tab Data Modifing \tab \code{\link{xpssRecode}} \tab Uses complete dataset\cr
\tab Data Exploring \tab \code{\link{xpssDescriptives}} \tab Uses working dataset only\cr
\tab Data Analyzing \tab \code{\link{xpssRegression}} \tab Uses working dataset only\cr
}
}
\examples{
data(fromXPSS)
# Filter Dataset by Variable V3 == 1
fromXPSS <- xpssFilter(x=fromXPSS, variable = "V3", filtervalue=1)
# Default Descriptive statistics for V6
xpssDescriptives(x=fromXPSS, variables = "V6")
# Turn Filter off
xpssFilterOff(x=fromXPSS)
}
\author{
Bastian Wiessner
}
\seealso{
Related Functions \code{\link{xpssDoIf}} \code{\link{xpssFilterOff}} \code{\link{xpssSample}} \code{\link{xpssSelectIf}} \code{\link{xpssTemporary}}
}

