% Generated by roxygen2 (4.0.2): do not edit by hand
\name{xpssFilterOff}
\alias{xpssFilterOff}
\title{Displays the content of variables.}
\usage{
xpssFilterOff(x)
}
\arguments{
\item{x}{a (non-empty) data.frame, data.table object or input data of class \code{xpssFrame}.}
}
\value{
If cases is not specified the return is the length of the data.
}
\description{
R Implementation of the SPSS \code{LIST} Function
}
\details{
LIST displays the content of selected variables. It's possible to display a sequenz with the \code{cases} argument.
}
\examples{
lala <- 1
}
\author{
Bastian Wiessner
}

