% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/xpssFlip.R
\name{xpssFlip}
\alias{xpssFlip}
\title{Flips variables}
\usage{
xpssFlip(x, variables = "all", names = NULL)
}
\arguments{
\item{x}{a (non-empty) data.frame, data.table object or input data of class "xpssFrame".}

\item{variables}{atomic character or character vector with the names of the variables to flip}

\item{names}{atomic character with the name of the variable for coloumn names.}
}
\value{
A flipped xpssFrame object. Returns the transpose of x.
}
\description{
R Implementation of the SPSS \code{FLIP} Function.
}
\examples{
data(fromXPSS)
xpssFlip(x=fromXPSS,variables=c("V4","V5","V6"),names="V1")
}
\author{
Bastian Wiessner
}
\seealso{
\code{\link{t}}
}

