% Generated by roxygen2 (4.0.2): do not edit by hand
\name{xpssFrame}
\alias{xpssFrame}
\title{Creates a xpssFrame Object}
\usage{
xpssFrame(x, as.data.table = FALSE, ...)
}
\arguments{
\item{x}{character string with the name of the file.}

\item{as.data.table}{a logical indicating whether you want to return the input data in data.table format.
Otherwise a data.frame returns. Default is FALSE.}

\item{\dots}{Arguments to pass on read.spss() from foreign.}
}
\description{
Creates numbers of new Variables
}
\details{
\code{x} the input data should be of the format \emph{.csv, .sav or .xlsx}.

The SPSS Variables are stored at the variable itself. A variable can have the following attributes:
value.labels, defined.MIS, MIS, varname, variable.label
}
\examples{
\dontrun{
daten <- xpssFrame("Testdata_1.sav")
}
}
\author{
Andreas Wygrabek
}
\seealso{
\code{\link{read.spss}}
}

