% Generated by roxygen2 (4.0.2): do not edit by hand
\name{xpssJDate}
\alias{xpssJDate}
\title{Days since 14. Oct 1582}
\usage{
xpssJDate()
}
\value{
Returns days since 14 Oct 1582
}
\description{
Count the number of days since October 14, 1582 (the first day of the Gregorian calendar).
}
\author{
Bastian Wiessner
}

