% Generated by roxygen2 (4.0.2): do not edit by hand
\name{xpssList}
\alias{xpssList}
\title{Displays the content of variables.}
\usage{
xpssList(x, variables = "", cases = FALSE,from = NULL, to = NULL, by = NULL)
}
\arguments{
\item{x}{a (non-empty) data.frame, data.table object or input data of class \code{xpssFrame}.}

\item{variables}{atomic character or character vector with the name of the variables.}

\item{cases}{Not documented yet}

\item{from}{Not documented yet}

\item{to}{Not documented yet}

\item{by}{Not documented yet}
}
\value{
If cases is not specified the return is the length of the data.
}
\description{
R Implementation of the SPSS \code{LIST} Function
}
\details{
LIST displays the content of selected variables. It's possible to display a sequenz with the \code{cases} argument.
}
\examples{
data(fromXPSS)

xpssList(x=fromXPSS)

xpssList(x=fromXPSS,
   variables = "V1")

xpssList(x=fromXPSS,
   variables = c("V1","V2"))

xpssList(x=fromXPSS,
   variables = span(fromXPSS,
                 from="V1",
                 to="V4"),
   cases =list(from=2,
               to=18,
               by=2))
}
\author{
Bastian Wiessner
}

