% Generated by roxygen2 (4.0.2): do not edit by hand
\name{xpssMissing}
\alias{xpssMissing}
\title{Displays if a variable contains missing values}
\usage{
xpssMissing (x, variable = NULL)
}
\arguments{
\item{x}{a (non-empty) data.frame, data.table object or input data of class "xpssFrame".}

\item{variable}{atomic character with the name of the variable.}
}
\value{
atomic logical with the length of the data. \cr Returns \code{TRUE},
if the variable contains a system- or user-defined missing value.
}
\description{
R Implementation of the SPSS \code{Missing} Function
}
\details{
Performs a missing value operation. \code{xpssMissing} displays potential system- or user-defined missing values of variables.
The different missing types will be inspected without checking if they are system-; or user-defined missings.

\code{variable} input, is limited to one variable.
}
\examples{
data(fromXPSS)
xpssMissing(fromXPSS, variable="V7_2")
}
\author{
Bastian Wiessner
}
\seealso{
Related Functions \code{\link{xpssNmiss}} , \code{\link{xpssNvalid}} , \code{\link{xpssSysmis}} ,\code{\link{xpssValue}}
}

