% Generated by roxygen2 (4.0.2): do not edit by hand
\name{xpssNmiss}
\alias{xpssNmiss}
\title{Displays the amount of missing values in variables.}
\usage{
xpssNmiss(x, variables = NULL)
}
\arguments{
\item{x}{a (non-empty) data.frame, data.table object or input data of class \code{"xpssFrame"}.}

\item{variables}{atomic character or character vector with the name of the variables.}
}
\value{
atomic numeric with the length of the data. Returns the amount of
system- or user-defined missing values of the variables.
}
\description{
R Implementation of the SPSS \code{Nmiss} Function
}
\details{
Performs a missing value operation. \code{xpssNmiss} displays the amount of system- or user-defined missing values of the variables. \cr User-defined and system-defind missings values get handled as one type of missing value.
}
\examples{
data(fromXPSS)
xpssNmiss(fromXPSS, variables=c("V6","V7_2"))
}
\author{
Bastian Wiessner
}
\seealso{
Related Functions \code{\link{xpssMissing}} , \code{\link{xpssNvalid}} , \code{\link{xpssSysmis}} ,\code{\link{xpssValue}}
}

