% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/xpssNOfCases.R
\name{xpssNofCases}
\alias{xpssNofCases}
\title{Selects the first n rows of the dataset}
\usage{
xpssNofCases(x, n = NULL)
}
\arguments{
\item{x}{input data.}

\item{n}{atomic numeric with the value of n.}
}
\value{
Output is the narrowed dataset.
}
\description{
R implementation of the SPSS \code{N OF CASES} argument.
}
\details{
\code{xpssNofCases} can be used to select via command the first N cases in the data file. xpssNofCases permanently modifies the data set.
 \strong{NOTE:} For temporary case selection, specify \code{\link{xpssTemporary}} before \code{xpssNofCases}.
}
\examples{
# load data
data(fromXPSS)
xpssNofCases(fromXPSS, n = 10)
}
\author{
Bastian Wiessner
}
\seealso{
Related Functions \code{\link{drop}} \code{\link{subset}}
}

