% Generated by roxygen2 (4.0.2): do not edit by hand
\name{xpssNvalid}
\alias{xpssNvalid}
\title{Displays the amount of valid values in variables.}
\usage{
xpssNvalid (x, variables = NULL)
}
\arguments{
\item{x}{a (non-empty) data.frame, data.table object or input data of class \code{"xpssFrame"}.}

\item{variables}{atomic character or character vector with the name of the variables.}
}
\value{
atomic numeric with the length of the data. Returns the amount of valid values of the variables.
}
\description{
R Implementation of the SPSS \code{Nvalid} Function
}
\details{
Performs a missing value operation. As opposite of \code{xpssNmiss}, \code{xpssNvalid} visualizies only the valid values of integer vectors or character vectors.

\code{variables} variables have to be same type.
}
\examples{
data(fromXPSS)
xpssNvalid(fromXPSS, variables=c("V6","V7_2"))
}
\author{
Bastian Wiessner
}
\seealso{
Related Functions \code{\link{xpssMissing}} , \code{\link{xpssNmiss}} , \code{\link{xpssSysmis}} ,\code{\link{xpssValue}}
}

