% Generated by roxygen2 (4.0.2): do not edit by hand
\name{xpssRenameVariables}
\alias{xpssRenameVariables}
\title{Renaming Variables}
\usage{
xpssRenameVariables(x, oldVarNames = NULL, newVarNames = NULL)
}
\arguments{
\item{x}{a (non-empty) data.frame, data.table object or input data of class \code{"xpssFrame"}.}

\item{oldVarNames}{Character vector naming the variables in the data set to rename.}

\item{newVarNames}{Character vector with the new variable names. Has to be the same length as oldVarNames.}
}
\value{
The functions returns the data x with new variable names.
}
\description{
xpssRenameVariables renames Variables within an exisiting data.frame or xpssFrame object.
}
\details{
Modifies names of one or more variables within a selected dataset.
}
\examples{
foo <- matrix(data = c(1:9),
nrow = 3,
ncol = 3,
dimnames = list(c(1,2,3),c("B","C","D")))
foo <- as.data.frame(foo)
foo <- as.xpssFrame(foo)
xpssRenameVariables(foo, c("B", "C", "D"), c("A", "B", "C"))
}
\author{
Andreas Wygrabek
}

