% Generated by roxygen2 (4.0.2): do not edit by hand
\name{xpssSample}
\alias{xpssSample}
\title{samples the a dataset}
\usage{
xpssSample(x, pct = NULL, n = NULL, from = NULL)
}
\arguments{
\item{x}{a (non-empty) data.frame, data.table object or input data of class \code{"xpssFrame"}.}

\item{pct}{Percentage to keep}

\item{n}{Number of cases to keep}

\item{from}{Basis for n}
}
\value{
a subset of the actual dataset. The subset get specified by pct or n.
}
\description{
Takes a sample from a xpssFrame, data frame or matrix
}
\details{
sample takes a sample of the specified size from the elements of x using either with or without replacement.

\code{pct} specifies a percentage of value which should be kept, allowed value range is from 1 to 0.01.
\code{n} indicates the amount of values to keep. \code{n} has to be lower then \code{from}.
\code{from} determines the basis for n. \code{from} has to be higher then \code{n}.
}
\examples{
data(fromXPSS)
xpssSample(fromXPSS, pct = 0.5)
}
\author{
Andreas Wygrabek
}
\seealso{
\code{\link{sample}}
}

