% Generated by roxygen2 (4.0.2): do not edit by hand
\name{xpssSelectIf}
\alias{xpssSelectIf}
\title{Creates a subset of cases}
\usage{
xpssSelectIf(x, cond = NULL)
}
\arguments{
\item{x}{a (non-empty) data.frame, data.table object or input data of class \code{"xpssFrame"}.}

\item{cond}{logical expression for subsetting the data}
}
\value{
Output is a subset of the actual dataset under the condition of the logical expression.
}
\description{
xpssSelectIf permanently selects cases for analysis based on logical conditions.
}
\details{
The conditions to select cases are specified in a logical expression. These logical expressions can contain relational operators, logical operators, arithmetic operations.

\strong{NOTE:} For temporary case selection, specify a TEMPORARY command before SELECT IF.
}
\examples{
#Select all Cases which match 1 in V3

data(fromXPSS)
temp <- xpssSelectIf(x=fromXPSS, cond = "V3 == 1")

temp <- xpssSelectIf(x=fromXPSS, cond="V4 == 1 & V7_1 < 200")
}
\author{
Andreas Wygrabek
}

