% Generated by roxygen2 (4.0.2): do not edit by hand
\name{xpssSortCases}
\alias{xpssSortCases}
\title{Sorts data ascending or descending}
\usage{
xpssSortCases(x, variables = NULL, order = "A")
}
\arguments{
\item{x}{a (non-empty) data.frame, data.table object or input data of class "xpssFrame".}

\item{variables}{atomic character or character vector with the name of the variables. Also "rownames" can be used to sort the data.}

\item{order}{either "A" for ascending order or "D" for descending order given as character strings. Has to be the same length as variables}
}
\value{
sorted data frame
}
\description{
xpssSort reorders the sequence of cases in the dataset based on the values of one or more variables.
}
\details{
Optionally the sorting can be specified in ascending or descending order for any variable. It is also possible to use combinations of ascending and descending order for different variables.
}
\examples{
data(fromXPSS)
xpssSortCases(fromXPSS, variables = c("V4", "V7_1", "V7_2"), order = c("A","D","A"))
}
\author{
Andreas Wygrabek
}
\seealso{
\code{\link{sort}} \code{link{order}}
}

