% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/xpssSplitFileOff.R
\name{xpssSplitFileOff}
\alias{xpssSplitFileOff}
\title{Set the data split off}
\usage{
xpssSplitFileOff(x)
}
\arguments{
\item{x}{a (non-empty) data.frame or input data of class \code{"xpssFrame"}.}
}
\value{
The function return the input data with modified attributes.
}
\description{
R implementation of the SPSS \code{SPLIT FILE} argument
}
\author{
Bastian Wiessner
}
\seealso{
\code{\link{xpssSplitFile}} \code{\link{xpssFilterOff}}
\code{\link{xpssTemporary}}
}

