% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/xpssTemporary.R
\name{xpssTemporary}
\alias{xpssTemporary}
\title{Temporary modification of the data}
\usage{
xpssTemporary(x)
}
\arguments{
\item{x}{a (non-empty) data.frame, data.table object or input data of class "xpssFrame".}
}
\description{
R implementation of the SPSS \code{TEMPORARY} Function. \code{xpssTemporary} modifies only the data for the following data-management procedure.
}
\details{
xpssTemporary signals the beginning of temporary transformation. Only the following data-management procedure takes affect on the data: \code{xpssRecode}, \code{xpssCount}, \code{xpssDoIf}, \code{xpssSelectIf}, \code{xpssSample}, \code{xpssFilter}, \code{xpssNumeric}, \code{xpssString}, \code{xpssVariableLabels}, \code{xpssValueLabels}, \code{xpssMissingValues}.
\cr\cr All the changes that are made are temporary. After the next modification the data is restored. \cr
For example:  all created variables, e.g. numeric or string variables created while \code{xpssTemporary} is in effect are temporary variables! \cr
Any changes or modifications made to existing variables while the \code{xpssTemporary} command is in effect are also \strong{temporary}! \cr
Any variables which are created or modified after this procedure are again permanent.

\tabular{rlll}{

\tab \code{Function Status} \tab Effect on created variables \tab Effect on modified variables \cr
\tab \code{Temporary ON} \tab Any created variable is temporary. \tab Any change on a variables is temporary. \cr
\tab \code{Temporary OFF} \tab Any created variable is permanent. \tab Any change on a variables is permanent.}

The xpssTemporary function allows analyses for subgroups without affecting the data and then repeat the analysis for the file as a whole.
}
\examples{
data(fromXPSS)
obj <- xpssTemporary(fromXPSS)
}
\author{
Andreas Wygrabek
}
\seealso{
Related Functions \code{\link{xpssDoIf}} \code{\link{xpssFilter}} \code{\link{xpssSample}} \code{\link{xpssSelectIf}}
}

