% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/xpssTime.R
\name{xpssTime}
\alias{xpssTime}
\title{Minutes since 14. Oct 1582}
\usage{
xpssTime()
}
\value{
Returns minutes since 14 Oct 1582
}
\description{
R implementation of the SPSS \code{$TIME} function. Count of the number of minutes since October 14, 1582 (the first day of the Gregorian calendar).
}
\author{
Bastian Wiessner
}

