% Generated by roxygen2 (4.0.2): do not edit by hand
\name{xpssValue}
\alias{xpssValue}
\title{Displays the value of the variable cases.}
\usage{
xpssValue(x, variables = NULL)
}
\arguments{
\item{x}{a (non-empty) data.frame, data.table object or input data of class "xpssFrame".}

\item{variables}{atomic character or character vector with the names of the variables.}
}
\value{
A "xpssFrame" object or (data frame) with implemented user-defined missing values.
}
\description{
R implementation of the SPSS \code{Value} function.
}
\details{
Performs a missing value operation. Adds the user-defined missing values of the variables in the existing dataset.

\code{variables} variables have to be same type.
}
\examples{
# load data
data(fromXPSS)

# add user defined values for variable V7_2
xpssValue(fromXPSS, variables="V7_2")

# add user defined values for variable V6 and V7_2
temp <- xpssValue(fromXPSS, variables=c("V6","V7_2"))
}
\author{
Bastian Wiessner
}
\seealso{
Related Functions \code{\link{xpssMissing}} , \code{\link{xpssNmiss}} , \code{\link{xpssNvalid}} , \code{\link{xpssSysmis}}
}

