% Generated by roxygen2 (4.0.2): do not edit by hand
\name{xpssVariableLabels}
\alias{xpssVariableLabels}
\title{Modifies variable labels}
\usage{
xpssVariableLabels(x, variables = NULL, labels = NULL)
}
\arguments{
\item{x}{a (non-empty) data.frame, data.table object or input data of class "xpssFrame".}

\item{variables}{Variable(s) to change the label. Given as a character string}

\item{labels}{Labels for the specified variables in  variables. Given as a character string. The labels are associated in order of appearence
to  variables.}
}
\value{
Input Data with modified attribute variable label
}
\description{
Changing the label of a variable. In the structure of xpss-data the variable label is an attribute of each variable.
}
\examples{
data(fromXPSS)
daten <- xpssVariableLabels(fromXPSS, c("V4", "V7_1"), c("Label1", "Label2"))
}
\author{
Andreas Wygrabek
}
\seealso{
\code{\link{attributes}} \code{link{attr}}
}

