% Generated by roxygen2 (4.0.2): do not edit by hand
\name{applyAttributeDemerge}
\alias{applyAttributeDemerge}
\title{Merges Attributdata}
\usage{
applyAttributeDemerge(x)
}
\arguments{
\item{x}{a (non-empty) data.frame, data.table object or input data of class \code{"xpssFrame"}.}
}
\value{
Output is the input data demerged on the basis of the attributes of the actual dataset.
}
\description{
Attribut merge function
}
\details{
Helper Function for DeMerging Attributedatasets
}
\author{
Bastian Wiessner
}
\keyword{internal}

