% Generated by roxygen2 (4.0.2): do not edit by hand
\name{applyAttributes}
\alias{applyAttributes}
\title{Apply stored attributes}
\usage{
applyAttributes(x, attributesToApply = NULL)
}
\arguments{
\item{x}{a data.frame, data.table object or input data of class xpssFrame.}

\item{attributesToApply}{to applied attributes}
}
\value{
Object with attributes from backup
}
\description{
Applies attributes stored by attributesBackup
}
\author{
Andreas Wygrabek
}
\keyword{internal}

