% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/applyMetaCheck.R
\name{applyMetaCheck}
\alias{applyMetaCheck}
\title{Apply Meta Checks}
\usage{
applyMetaCheck(x, pos = 1, envir = as.environment(pos))
}
\arguments{
\item{x}{a (non-empty) data.frame, data.table object or input data of class \code{"xpssFrame"}.}
}
\value{
Output is the input data merged on the basis of the condition in the attributes of the actual dataset.
}
\description{
Checks the attributes for eventual merging the data
}
\details{
Helper Function for Merging Attributedatasets
}
\author{
Bastian Wiessner
}
\keyword{internal}

