% Generated by roxygen2 (4.0.2): do not edit by hand
\name{as.xpssFrame}
\alias{as.xpssFrame}
\title{Coerces data.frames to xpssFrame objects}
\usage{
as.xpssFrame(x)
}
\arguments{
\item{x}{any (non-empty) R object.}
}
\value{
\code{as.xpssFrame} returns an xpssFrame, with the essential attributes \code{DO_IF}, \code{FILTER}, \code{SPLIT_FILE}, \code{TEMPORORY} and \code{WEIGHTS} for the actual dataset and additional variable.label and varname attributes for every variable in the dataset.
\cr \cr \code{is.xpssFrame} returns TRUE if its argument is a xpssFrame (that is, has "xpssFrame" amongst its classes) and FALSE otherwise.
}
\description{
Function to check if an object is an xpssFrame, or coerce it to one if possible.
}
\examples{
# example data.frame
temp <- data.frame(x=1:5, y = 2:6, z=c("a","b","c","d","e"))
# coerce to xpssFrame
temp <- as.xpssFrame(temp)
# check if it is an xpssFrame
is.xpssFrame(temp)
}
\author{
Bastian Wiessner
}

