% Generated by roxygen2 (4.0.2): do not edit by hand
\name{attributesBackup}
\alias{attributesBackup}
\title{Stores Attributes}
\usage{
attributesBackup(x)
}
\arguments{
\item{x}{a (non-empty) data.frame, data.table object or input data of class \code{"xpssFrame"}.}

\item{cond}{logical expression for subsetting the data}
}
\value{
Output is a subset of the actual dataset under the condiftion of the logical expression.
}
\description{
Attribut backup function
}
\details{
The conditions to select cases are specified in a logical expression. These logical expressions can contain relational operators, logical operators, arithmetic operations.

\strong{NOTE:} For temporary case selection, specify a TEMPORARY command before SELECT IF.
}
\examples{
data(fromXPSS)
data <- xpssSelectIf(fromXPSS, cond = "V3 == 1")
}
\author{
Bastian Wiessner
}
\keyword{internal}

