% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/xpssComputeArithmetics.R
\name{computeArsin}
\alias{computeArsin}
\title{Computes the arc-sine.}
\usage{
computeArsin(x)
}
\arguments{
\item{x}{an atomic numeric or numeric vector or numeric matrix.}
}
\value{
returns a vector with the arc-sine values of the input data.
}
\description{
Helper Function for xpssCompute. R Implementation of the SPSS \code{ARSIN} Function.
}
\examples{
# atomic numeric input with values between -1 and 1
x1 <- rnorm(n=20,mean=0,sd=0.25)
xpssCompute(x=x1,fun="computeArsin")
# numeric vector input with values between -1 and 1
x2 <- c(rnorm(n=20,mean=0,sd=0.25),rnorm(n=20,mean=0,sd=0.35))
xpssCompute(x=x2,fun="computeArsin")
# numeric matrix input with values between -1 and 1
x3 <- rbind(rnorm(n=20,mean=0,sd=0.25),rnorm(n=20,mean=0,sd=0.35),rnorm(n=20,mean=0,sd=0.5))
xpssCompute(x=x3,fun="computeArsin")
}
\author{
Bastian Wiessner
}
\seealso{
\code{\link{asin}}
}
\keyword{internal}

