% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/xpssComputeStrings.R
\name{computeChar_lpad}
\alias{computeChar_lpad}
\title{Expand strings on the left}
\usage{
computeChar_lpad (x, length = NULL, fill = NULL)
}
\arguments{
\item{x}{character or character vector.}

\item{length}{numeric. Number of characters x is to be filled on the left.}

\item{fill}{optional. String which x should be filled with.}
}
\value{
String, left-padded by length \code{length}.
}
\description{
Helper Function for xpssCompute. R Implementation of the SPSS \code{CHAR.LPAD} Function.
}
\details{
The value of length represents the number of characters and must be a positive integer. If the optional argument \code{fill} is not specified, \code{x} is padded with blank spaces.
}
\examples{
computeChar_lpad("My Friend",15)
computeChar_lpad("My Friend",15,"Hello ") # Muss noch angepasst werden, wenn Funktion fertig ist
}
\author{
Bastian Wiessner
}
\seealso{
\code{\link{paste0}}
}
\keyword{internal}

