% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/xpssComputeStrings.R
\name{computeConcat}
\alias{computeConcat}
\title{computeConcat (x, sep = "")}
\usage{
computeConcat (x, sep = "")
}
\arguments{
\item{x}{character or character vector.}

\item{sep}{atomic character element which splits the input data. Default is "".}
}
\value{
String. Returns a string that is the concatenation of all its arguments.
}
\description{
Helper Function for xpssCompute. R Implementation of the SPSS \code{CONCAT} Function.
}
\examples{
x <- c("this","is","a","vector")
computeConcat(x)
}
\author{
Bastian Wiessner
}
\seealso{
\code{\link{paste0}}
}
\keyword{internal}

