% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/xpssComputeArithmetics.R
\name{computeMod}
\alias{computeMod}
\title{Returns the remainder of a division.}
\usage{
computeMod(x,modulus = NULL)
}
\arguments{
\item{x}{an atomic numeric or numeric vector or numeric matrix.}

\item{modulus}{atomic numeric \code{x} is divided by.}
}
\value{
returns a numeric vector including the remainder of \code{x} by \code{modulus}.
}
\description{
Helper Function for xpssCompute. R Implementation of the SPSS \code{MOD} Function.
}
\examples{
# atomic numeric input
x1 <- seq(from = 1,to = 20,by = 1.5)
xpssCompute(x = x1,fun = "computeMod", modulus = 2)
# numeric vector input
x2 <- c(seq(from = 1,to = 20,by = 1.5),seq(from = 1,to = 20,by = 1.25))
xpssCompute(x = x2,fun = "computeMod", modulus = 3)
# numeric matrix input
x3 <- rbind(seq(from = 1,to = 10,by = 1),seq(from = 1,to = 20,by = 2),seq(from = 1,to = 30,by = 3))
xpssCompute(x = x3,fun = "computeMod", modulus = 4)
}
\author{
Bastian Wiessner
}
\seealso{
\code{\link{\%\%}}
}
\keyword{internal}

