% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/xpssComputeStrings.R
\name{computeReplace}
\alias{computeReplace}
\title{Replace matched pattern in a string}
\usage{
computeReplace (x,pattern=NULL,match=NULL, count = NULL)
}
\arguments{
\item{x}{input vector}

\item{pattern}{pattern to look for}

\item{match}{string tp replace with}

\item{count}{number of occurencees of pattern to replace}
}
\value{
String. Returns a character vector
}
\description{
Helper Function for xpssCompute. R Implementation of the SPSS \code{REPLACE} Function.
}
\examples{
computeReplace("Makfts sftnsft","ft","e")
}
\author{
Bastian Wiessner
}
\seealso{
\code{\link{str_replace}}\code{\link{str_replace_all}}\code{\link{grep}}
}
\keyword{internal}

