% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/xpssComputeArithmetics.R
\name{computeRnd}
\alias{computeRnd}
\title{Rounds values.}
\usage{
computeRnd(x, digits=2)
}
\arguments{
\item{x}{a numeric or complex vector.}

\item{digits}{atomic numeric argument, indicating the number of decimal places.}
}
\value{
returns the rounded values of \code{x} with the specified number of decimal places.
}
\description{
Helper Function for xpssCompute. R Implementation of the SPSS \code{RND} Function.
}
\examples{
# atomic numeric input
xpssCompute(x = fromXPSS, variables = "V5",fun = "computeRnd")

xpssCompute(x = fromXPSS, variables = "V5",fun = "computeRnd",digits=1)
}
\author{
Bastian Wiessner
}
\seealso{
\code{\link{round}}
}
\keyword{internal}

