% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/xpssComputeArithmetics.R
\name{computeSqrt}
\alias{computeSqrt}
\title{Computes the square root}
\usage{
computeSqrt(x)
}
\arguments{
\item{x}{an atomic numeric or numeric vector or numeric matrix.}
}
\value{
returns a numeric, numeric vector or matrix with the square root values of the input data.
}
\description{
Helper Function for xpssCompute. R Implementation of the SPSS \code{SQRT} Function.
}
\examples{
# atomic numeric input
xpssCompute(x = fromXPSS, variables = "V5",fun = "computeSqrt")
}
\author{
Bastian Wiessner
}
\seealso{
\code{\link{sqrt}}
}
\keyword{internal}

