% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/xpssComputeDates.R
\name{computeXdate_mday}
\alias{computeXdate_mday}
\title{Extracts the date of month on basis of a given date}
\usage{
computeXdate_mday(x=NULL)
}
\arguments{
\item{x}{atomic object of class \code{character, POSIXlt  or  POSIXt} holding date data.}
}
\value{
Returns a character string with the date's day.
}
\description{
R Implementation of the SPSS \code{XDATE.MDAY} Function. \code{computeXdate_mday} is a helper function for xpssCompute.
}
\details{
computeXdate_mday extract the day component out of the date string, the other date and time componentes will be omitted.
}
\examples{
# 046
computeXdate_mday("2015-02-15 20:22:20")
# 293
computeXdate_mday("2022-10-20 21:22:12")
}
\author{
Bastian Wiessner
}
\seealso{
\code{\link{computeXdate_date}} \code{\link{computeDate_wkyr}}
}
\keyword{internal}

