% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/xpssComputeDates.R
\name{computeXdate_minute}
\alias{computeXdate_minute}
\title{Extracts the minute component on basis of a given date}
\usage{
computeXdate_minute(x=NULL)
}
\arguments{
\item{x}{atomic object of class \code{character, POSIXlt  or  POSIXt} holding date data.}
}
\value{
Returns a character string with the minute component.
}
\description{
R Implementation of the SPSS \code{XDATE.MINUTE} Function. \code{computeXdate_minute} is a helper function for xpssCompute.
}
\details{
computeXdate_minute extract the minute component out of the date string, the other date and time componentes will be omitted.
}
\examples{
# 22
xpssCompute(x="2022-10-20 21:22:12", fun="computeXdate_minute")
# 11
xpssCompute(x="2015-02-15 20:11:20", fun="computeXdate_minute")
}
\author{
Bastian Wiessner
}
\seealso{
\code{\link{computeXdate_date}} \code{\link{computeDate_wkyr}}
}
\keyword{internal}

