% Generated by roxygen2 (4.0.2): do not edit by hand
\name{span}
\alias{span}
\title{Indicates range of varlist}
\usage{
span(x, from = NULL, to = NULL, addDF = FALSE)
}
\arguments{
\item{x}{a (non-empty) data.frame, data.table object or input data of class "xpssFrame".}

\item{from}{variable that opens the span}

\item{to}{variable that closes the span}

\item{addDF}{Should the name of the input data be used?}
}
\value{
Returns a varlist with the name of the variables which a within the range of the span indicator.
}
\description{
Creates a list of variables within a specific range
}
\examples{
foo <- matrix(data = c(1:15),
nrow = 3,
ncol = 5,
dimnames = list(c(1,2,3),c("A","B","C","D","E")))
span(foo, from = "C", to = "E",addDF = FALSE)
span(foo, from = "C", to = "E",addDF = TRUE)
}
\author{
Andreas Wygrabek
}

