% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/span.R
\name{span}
\alias{span}
\title{Indicates range of varlist}
\usage{
span(x, from = NULL, to = NULL, addDF = FALSE)
}
\arguments{
\item{x}{a (non-empty) data.frame or input data of class "xpssFrame".}

\item{from}{the variable that opens the span.}

\item{to}{the variable that closes the span.}

\item{addDF}{if the name of the input data should be used?}
}
\value{
Returns a varlist with the name of the variables which are within the range of the span indicator.
}
\description{
Creates a list of variables within a specific range.
}
\examples{
# load data
data(fromXPSS)
# span from V1 to V5
span(x=fromXPSS,from="V1",to="V5",addDF=FALSE)
# span from V3 to V1 and add the name of the data object.
span(x=fromXPSS,from="V3",to="V1",addDF=TRUE)
}
\author{
Andreas Wygrabek
}

