% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/translate2R.R
\docType{package}
\name{translateSPSS2R-package}
\alias{translateSPSS2R-package}
\title{Dokumentation}
\arguments{
\item{Package:}{translateSPSS2R}

\item{Type:}{Package}

\item{Version:}{1.0}

\item{Date:}{11.12.2014}

\item{Imports:}{car ,data.table, e1071, foreign, lattice, plyr, stringr, tidyr, zoo}

\item{License:}{GPL-3}

\item{Maintainer:}{Andreas Wygrabek}
}
\description{
translateSPSS2R delivers a function set for translated commands of SPSS. The usage is oriented on the handling of the SPSS-Syntax.
}
\details{
translateSPSS2R delivers a function set for translated commands of SPSS. The usage is oriented on the handling of the SPSS-Syntax. Mainly the package has two purposes.
\enumerate{
\item It facilitates SPSS-Users to change over to R.
\item It facilitates migration projects from SPSS to R.
}
}
\author{
Andreas Wygrabek <Andreas.Wygrabek@eoda.de> and Bastian Wiessner <Bastian.Wiessner@eoda.de>
}
\references{
\url{https://service.eoda.de/translater/?lang=en}
\cr \url{http://www.eoda.de/en}
}

