% Generated by roxygen2 (4.0.2): do not edit by hand
\name{xpssAddValueLabels}
\alias{xpssAddValueLabels}
\title{Modifies value labels}
\usage{
xpssAddValueLabels(x, variables = NULL, values = NULL, labels = NULL,
  datevariables = NULL, datevalues = NULL, datelabels = NULL)
}
\arguments{
\item{x}{a (non-empty) data.frame or input data of class \code{"xpssFrame"}.}

\item{variables}{atomic character or character vector with the name of the variables.}

\item{values}{atomic numeric or numeric vector containing the value of the variable.}

\item{labels}{character string containing a variable's label.}

\item{datevariables}{atomic date or date vector with the name of the date variables.}

\item{datevalues}{atomic date or date vector containing the value of the date, the value has to be like the old date format.}

\item{datelabels}{character string containing a date's label.}
}
\description{
xpssAddValueLabels appends value labels for specific variables. The values of the label get stored in attributes of the variable.
}
\details{
The values labels are stored in the variable itself.
\cr In contrast to \code{\link{xpssValueLabels}} , \code{\link{xpssAddValueLabels}} does not erase existing value labels. \cr If the value label for a specific variable already exists, this value label gets overwritten. \cr If the value label for a specific variable does not exist, the value label gets created, without deleting the existing value labels for that variable.
}
\examples{
data(fromXPSS)

temp <- xpssValueLabels(fromXPSS,
                           variables = "V1",
                           value = 1 ,
                           label = "Label1")

temp <- xpssAddValueLabels(temp,
                           variables = "V1",
                           value = 2 ,
                           label = "Label2")

attributes(temp$V1)$value.labels
}
\author{
Bastian Wiessner
}
\seealso{
\code{\link{read.spss}} \code{\link{xpssValueLabels}} \code{\link{xpssVariableLabels}}
}

