% Generated by roxygen2 (4.0.2): do not edit by hand
\name{xpssAny}
\alias{xpssAny}
\title{Selecting Cases by Condition}
\usage{
xpssAny(x, st = NULL, nd = NULL)
}
\arguments{
\item{x}{a (non-empty) data.frame or input data of class \code{"xpssFrame"}.}

\item{st}{Value to search for OR variable where to search in.}

\item{nd}{Values to search for OR variables to search in.}
}
\value{
The functions returns a logical vector.
}
\description{
xpssAny can be perceived as a wrapper function for \%in\% applicable on more than one variable.
}
\examples{
data(fromXPSS)
xpssAny(fromXPSS, 310, c("V7_1", "V7_2"))
xpssAny(fromXPSS, "V7_1", c(310,320,170))

xpssAny(fromXPSS, "Audi", c("V1", "V7_2"))
xpssAny(fromXPSS, "V1", c("Audi"))
}
\author{
Andreas Wygrabek
}

