% Generated by roxygen2 (4.0.2): do not edit by hand
\name{xpssAny}
\alias{xpssAny}
\title{Selecting cases by condition}
\usage{
xpssAny(x, st = NULL, nd = NULL)
}
\arguments{
\item{x}{a (non-empty) data.frame or input data of class \code{"xpssFrame"}.}

\item{st}{atomic numeric or atomic character with a single value to search for OR variable where to search in.}

\item{nd}{atomic numeric or atomic character, respecetively as numeric vector or character vector with values to search for OR variables to search in.}
}
\value{
A logical vector with matched conditions.
}
\description{
xpssAny can be perceived as a wrapper function for \code{\link{\%in\%}} applicable on more than one variable.
}
\examples{
# load data
data(fromXPSS)

# search 310 in V7_1 and V7_2
xpssAny(fromXPSS, 310, c("V7_1", "V7_2"))

# check whether V7_1 contains 310 or 320 or 170
xpssAny(fromXPSS, "V7_1", c(310,320,170))

# search for Auto in V1 and V7_2
xpssAny(fromXPSS, "Audi", c("V1", "V7_2"))

# check whether V1 contains Audi
xpssAny(fromXPSS, "V1", c("Audi"))
}
\author{
Andreas Wygrabek
}
\seealso{
\code{\link{xpssCount}} \code{\link{\%in\%}} \code{\link{is.element}}
}

