% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/xpssCasenum.R
\name{xpssCasenum}
\alias{xpssCasenum}
\title{Sequential numbering of cases}
\usage{
xpssCasenum(x)
}
\arguments{
\item{x}{a (non-empty) data.frame or input data of class \code{"xpssFrame"}.}
}
\value{
Returns a sequential atomic numeric or numeric vector.
}
\description{
R implementation of the SPSS \code{$CASENUM} system variable. xpssCasenum counts the number of cases within a variable, or respectively the number of observations in the dataset.
}
\details{
xpssCasenum fit well as ID-Variable.
}
\examples{
# load data
data(fromXPSS)

# create id variable
fromXPSS$id <- xpssCasenum(fromXPSS)
}
\author{
Bastian Wiessner
}
\seealso{
Related Functions \code{\link{xpssCount}} , \code{\link{xpssDate}} , \code{\link{xpssDate11}}
}

