% Generated by roxygen2 (4.0.2): do not edit by hand
\name{xpssCount}
\alias{xpssCount}
\title{Counts amount of specific Observations}
\usage{
xpssCount(x, variables = NULL, count = NULL)
}
\arguments{
\item{x}{a (non-empty) data.frame, data.table object or input data of class \code{xpssFrame}.}

\item{variables}{atomic character or character vector with the name of the variables.}

\item{count}{value to count}
}
\value{
A vector of the same length as x.
}
\description{
R Implementation of the SPSS \code{$COUNT} system variable
}
\details{
Count displays the amount observations which match the count statement.
}
\examples{
data(fromXPSS)

xpssCount(x=fromXPSS,
   variables = "V1", count="Nissan")

xpssCount(x=fromXPSS,
   variables = "V5", count=2)

xpssCount(x=fromXPSS,
   variables = c("V5","V5_kl2"), count=2)
}
\author{
Bastian Wiessner
}
\seealso{
Related Functions \code{\link{xpssAny}}
}

