% Generated by roxygen2 (4.0.2): do not edit by hand
\name{xpssCount}
\alias{xpssCount}
\title{Counts frequencies of specific observations}
\usage{
xpssCount(x, variables = NULL, count = NULL)
}
\arguments{
\item{x}{a (non-empty) data.frame, data.table object or input data of class \code{xpssFrame}.}

\item{variables}{atomic character or character vector with the name of the variables.}

\item{count}{atomic character or atomic numeric pattern.}
}
\value{
A vector of the same length as x.
}
\description{
R implementation of the SPSS \code{$COUNT} system variable.
}
\details{
Count displays the frequencies of observations matching the count statement.
}
\examples{
# load data
data(fromXPSS)

# count Nissan in variable V1
xpssCount(x=fromXPSS,
   variables = "V1", count=list(exact="Nissan"))

# count 2 in variable V5
xpssCount(x=fromXPSS,
   variables = "V5", count=list(exact=2))

# count in V5 to V7_2, lowest values until 100
xpssCount(fromXPSS,
   variables = span(fromXPSS, from = "V5",
                             to = "V7_2"),
                   count = list(from = "lo",
                                  to = 100))
}
\author{
Bastian Wiessner
}
\seealso{
Related Functions \code{\link{xpssAny}} \code{\link{\%in\%}} \code{\link{is.element}}
}

