% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/xpssDate.R
\name{xpssDate}
\alias{xpssDate}
\title{Today's date}
\usage{
xpssDate()
}
\value{
Current date as "dd-mmm-yy". Day and year are numeric values with the length two, month is a character string with the length three.
}
\description{
R implementation of the SPSS \code{$date} system variable.
}
\details{
xpssDate provides the SPSS format of dates.
}
\examples{
xpssDate()
}
\author{
Bastian Wiessner
}

