\name{xpssDate11}
\alias{xpssDate11}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ 
Date function with four-digit-year.
}
\description{
R Implementation of the SPSS \code{$DATE11} system variable
}
\usage{
xpssDate11()
}
%- maybe also 'usage' for other objects documented here.
\details{
Current date in international date format with four-digit year
}
\value{
%%  ~Describe the value returned
%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
Bastian Wiessner
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
Related Functions \code{\link{xpssTime}} \code{\link{xpssDate}} code{\link{xpssJDate}}
}
\examples{
xpssDate11()
}