% Generated by roxygen2 (4.0.2): do not edit by hand
\name{xpssEndIf}
\alias{xpssEndIf}
\title{Creates a subset of cases}
\usage{
xpssEndIf(x)
}
\arguments{
\item{x}{a (non-empty) data.frame or input data of class \code{"xpssFrame"}.}
}
\value{
Output is a  the original dataset.
}
\description{
xpssEndIf determines the end of the analysis based on logical conditions via xpssDoIf.
}
\details{
\code{\link{xpssEndIf}} restores the original data, after \code{\link{xpssDoIf}} subsetted the data. All changes which were made until \code{\link{xpssEndIf}} will be taken over, the excluded data will be untouched!\cr \cr

\strong{NOTE:} For temporary case selection, specify \code{\link{xpssTemporary}} before \code{\link{xpssDoIf}}.
}
\examples{
data(fromXPSS)
# Select all cases which match 1 in V3
temp <- xpssDoIf(x=fromXPSS, cond = "V3 == 1")
# Recode all selected cases
temp <- xpssRecode(x=temp,varin="V5",rec="lo:78 = 1; else = 2")
# End DoIf Subsetting, restore dataset with applied changes
temp <- xpssEndIf(x=temp)
}
\author{
Andreas Wygrabek
}

