% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/xpssJDate.R
\name{xpssJDate}
\alias{xpssJDate}
\title{Days since 14. Oct 1582}
\usage{
xpssJDate()
}
\value{
Returns days since 14 Oct 1582
}
\description{
Count the number of days since October 14, 1582.
}
\examples{
xpssJDate()
}
\author{
Bastian Wiessner
}

