% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/xpssNOfCases.R
\name{xpssNOfCases}
\alias{xpssNOfCases}
\title{Selects the first n rows of the dataset}
\usage{
xpssNOfCases(x, n = NULL)
}
\arguments{
\item{x}{input data.}

\item{n}{atomic numeric with the value of n.}
}
\value{
Output is the narrowed dataset.
}
\description{
R implementation of the SPSS \code{N OF CASES} argument.
}
\examples{
# load data
data(fromXPSS)
xpssNOfCases(fromXPSS, n = 10)
}
\author{
Bastian Wiessner
}
\seealso{
Related Functions \code{\link{drop}} \code{\link{subset}}
}

