% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/xpssSample.R
\name{xpssSample}
\alias{xpssSample}
\title{Creates a sample}
\usage{
xpssSample(x, pct = NULL, n = NULL, from = NULL)
}
\arguments{
\item{x}{a (non-empty) data.frame, data.table object or input data of class \code{"xpssFrame"}.}

\item{pct}{atomic numeric, determines the percentage to keep.}

\item{n}{atomic numeric, specifies the number of cases to keep.}

\item{from}{atomic numeric, indicates the Basis for n.}
}
\value{
Returns a subset of the actual dataset.
}
\description{
R mplementation of the SPSS \code{SAMPLE} argument. Takes a sample from a xpssFrame object, data frame or matrix.
}
\details{
xpssSample takes a sample of the specified size from the elements of x either with or without replacement. The subset get specified by pct or n.

\code{pct} specifies a percentage value, for the amount of data which should be kept, allowed value range is from 0.01 to 1.
\code{n} indicates the amount of values to keep. \code{n} has to be lower then \code{from}.
\code{from} determines the basis for n. \code{from} has to be higher then \code{n}.
}
\examples{
# load data
data(fromXPSS)

# draw a sample, with 50\% percent of the original data
xpssSample(fromXPSS, pct = 0.5)
}
\author{
Andreas Wygrabek
}
\seealso{
\code{\link{sample}}
}

