% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/xpssSortVariables.R
\name{xpssSortVariables}
\alias{xpssSortVariables}
\title{Sorting variables}
\usage{
xpssSortVariables(x, by = NULL, order = "A")
}
\arguments{
\item{x}{dataframe to sort}

\item{by}{argument to sort by}

\item{order}{Up or Down / A or D}
}
\value{
Returns \code{x} sorted by the \code{by} argument.
}
\description{
R implementation of the SPSS \code{SORT VARIABLES} function.
}
\details{
\code{x} can be sorted based on the following \code{by} arguments.
 \itemize{
 \item \emph{NAME.} Sort by variable names. Primary sort is alphabetical. Trailing digits are sorted numerically
 \item \emph{TYPE.} Sort variables by type (numeric or string). Sort string variables by width.
 \item \emph{FORMAT.} Sort variables by format (Date, String, ...).
 \item \emph{LABEL.} Sort variables alphabetical by variable labels.
 \item \emph{VALUES.} Sort variables by value lables.
 \item \emph{MISSING.} Sort variables by missing values.
 \item \emph{MEASURE.} Sort variables by scale level (Order: nominal, ordinal, scale).
 \item \emph{ROLE.} Sort variables by role (For example: input, goal, both, non).
 \item \emph{COLUMNS.} Sort variables by column width.
 \item \emph{ALIGNMENT.} Sort variables by alignment (left, right, center).
 \item \emph{ATTRIBUTE name.} Sort variables by values of the specified custom variable attribute name.
 }
}
\examples{
data(fromXPSS)

# Returns variables in alphabetical order (A-Z).
xpssSortVariables(fromXPSS, by = "NAME")

# Returns variables in alphabetical order (Z-A).
xpssSortVariables(fromXPSS, by = "NAME", order = "D")

# Same output as above.
xpssSortVariables(fromXPSS, by = "NAME", order = "D")

# Returns variables ordered by type (numeric-string).
# Shorter string variables come before longer string variables.
xpssSortVariables(fromXPSS, by = "TYPE", order = "A")

# Returns variables ordered by type (string-numeric).
# Longer string variables come before shorter string variables.
xpssSortVariables(fromXPSS, by = "TYPE", order = "D")

# Returns variables ordered by column width (short-long)
xpssSortVariables(fromXPSS, by = "COLUMNS", order = "A")

# Returns variables ordered by column width (long-short)
xpssSortVariables(fromXPSS, by = "COLUMNS", order = "D")
}
\author{
Benjamin Piest
}
\seealso{
\code{\link{sort}}
}

