% Generated by roxygen2 (4.0.2): do not edit by hand
\name{xpssSplitFile}
\alias{xpssSplitFile}
\title{Splits the data.frame into pieces by a factor}
\usage{
xpssSplitFile(x, by = NULL, type = "seperated", attachSplit = FALSE)
}
\arguments{
\item{x}{a (non-empty) data.frame, data.table object or input data of class \code{"xpssFrame"}.}

\item{by}{factor variable which splits the data frame.}

\item{type}{character vector containing, either seperated or layered. Specifies the results from an analysis function applied on a splitted data frame. Default is \code{"seperated"} design.}

\item{attachSplit}{logical. Indicating if the splitted data should get attached to the object as an attribute. Default is \code{"FALSE"}.}
}
\value{
The function return the input data with modified attributes.
}
\description{
R implementation of the SPSS \code{SPLIT FILE} argument
}
\author{
Andreas Wygrabek
}
\seealso{
\code{\link{xpssDoIf}} \code{\link{xpssFilterOff}} \code{\link{xpssSelectIf}}
\code{\link{xpssTemporary}}
}

