% Generated by roxygen2 (4.0.2): do not edit by hand
\name{xpssSplitFile}
\alias{xpssSplitFile}
\title{Splits the data.frame into pieces by a factor}
\usage{
xpssSplitFile(x, by = NULL, type = "seperated", attachSplit = FALSE)
}
\arguments{
\item{x}{a (non-empty) data.frame, data.table object or input data of class \code{"xpssFrame"}.}

\item{by}{the factor variable which splits the data frame}

\item{type}{the results from an analysis function applied on a splitted data frame can be presented either in a seperated or an layered design}

\item{attachSplit}{Should the splitted data be attached to the object as an attribute?}
}
\value{
The function return the input data with modified attributes
}
\description{
R Implementation of the SPSS \code{filter} argument
}
\examples{
data(fromXPSS)
xpssSplitFile(fromXPSS, "V6", "layered", attachSplit = FALSE)
}
\author{
Andreas Wygrabek
}
\seealso{
\code{\link{xpssDoIf}} \code{\link{xpssFilterOff}} \code{\link{xpssSelectIf}}
\code{\link{xpssTemporary}}
}

