% Generated by roxygen2 (4.0.2): do not edit by hand
\name{xpssTemporary}
\alias{xpssTemporary}
\title{Temporary modification of the data}
\usage{
xpssTemporary(x)
}
\arguments{
\item{x}{a (non-empty) data.frame, data.table object or input data of class "xpssFrame".}
}
\description{
xpssTemporary modifies only the data for the following data-management procedure.
}
\details{
xpssTemporary signals the beginning of temporary transformation. Only the following data-management procedure takes affect on the data. All the changes that are made are temporary. After the next modification the data is restored. \cr
For example:  all created variables, e.g. numeric or string variables created while the TEMPORARY is in effect are temporary variables! \cr
Any changes or modifications made to existing variables while the TEMPORARY command is in effect are also temporary! \cr
Any variables which are created or modified after this procedure are again permanent. \cr \cr

The xpssTemporary Function allows analyses for subgroups without affecting the data and then repeat the analysis for the file as a whole.
}
\examples{
data(fromXPSS)
obj <- xpssTemporary(fromXPSS)
}
\author{
Andreas Wygrabek
}

