% Generated by roxygen2 (4.0.2): do not edit by hand
\name{xpssValue}
\alias{xpssValue}
\title{Displays the value of the variable cases.}
\usage{
xpssValue(x, variables = NULL)
}
\arguments{
\item{x}{a (non-empty) data.frame, data.table object or input data of class "xpssFrame".}

\item{variables}{atomic character or character vector with the name of the variables.}
}
\value{
A "xpssFrame" object or (data frame) with implemented user-defined missing values.
}
\description{
R Implementation of the SPSS \code{Value} Function
}
\details{
Performs a missing value operation. Adds the user-defined missing values of the variables in the existing dataset.

\code{variables} variables have to be same type.
}
\examples{
data(fromXPSS)
temp <- xpssValue(fromXPSS, variables="V7_2")

temp <- xpssValue(fromXPSS, variables=c("V6","V7_2"))
}
\author{
Bastian Wiessner
}
\seealso{
Related Functions \code{\link{xpssMissing}} , \code{\link{xpssNmiss}} , \code{\link{xpssNvalid}} , \code{\link{xpssSysmis}}
}

