% Generated by roxygen2 (4.0.2): do not edit by hand
\name{xpssVarsToCases}
\alias{xpssVarsToCases}
\title{Indicates range of varlist}
\usage{
xpssVarsToCases(x, from, idVar = NULL, indexVar = NULL, nullArg = "keep", countVar = NULL, varLabels = list(id = NULL, index = NULL, count = NULL))
}
\arguments{
\item{x}{a (non-empty) data.frame, data.table object or input data of class "xpssFrame".}

\item{from}{variable that opens the span}

\item{idVar}{Should an id-Variable be created?}

\item{indexVar}{Should an index-Variable be created?}

\item{nullArg}{Can be either "keep" or "drop"}

\item{countVar}{Should a counter be created?}

\item{varLabels}{Should be labels for the id-, index- and count variable setted?}
}
\value{
Returns the transformed xpssFrame
}
\description{
Creates a list of variables within a specific range
}
\examples{
data(fromXPSS)
xpssVarsToCases(fromXPSS, from = list(c("newVar", "V7_1, V7_2")), idVar = "myID", indexVar = "myIndex", nullArg = "drop", countVar = "Counter")
}
\author{
Andreas Wygrabek
}

